Move geometry related classes to a new module
