+++ svg_timeline/timeline_elements.py
@@ -5,74 +5,13 @@ from pathlib import Path
 from typing import Optional
 
 from svg_timeline.geometry import Vector
-from svg_timeline.style import GeometrySettings, ClassNames
+from svg_timeline.style import ClassNames
 from svg_timeline.svg import SvgGroup
 from svg_timeline.svg_primitives import Line, Text, Circle, Image, Rectangle
-from svg_timeline.time_calculations import TimeGradient, TimeSpacing
-
+from svg_timeline.time_calculations import TimeSpacing
+from svg_timeline.timeline_geometry import TimeLineGeometry, GeometrySettings
 
 Classes = Optional[list[str]]
-
-
-class TimeLineGeometry:
-    """ class for the transfer of dates and lanes to canvas coordinates """
-    def __init__(self,
-                 start_date: datetime,
-                 end_date: datetime,
-                 style: Optional[GeometrySettings] = None,
-                 ):
-        """
-        :param start_date: the lower boundary of the timeline
-        :param end_date: the upper boundary of the timeline
-        """
-        self._style = style or GeometrySettings()
-        self._first = start_date
-        self._last = end_date
-        y = self._style.lane.lane_zero_y * self._style.canvas.height
-        x1 = self._style.canvas.x_padding * self._style.canvas.width
-        x2 = (1 - self._style.canvas.x_padding) * self._style.canvas.width
-        self._gradient = TimeGradient(source=Vector(x1, y), target=Vector(x2, y),
-                                      start_date=start_date, end_date=end_date)
-
-    @property
-    def style(self) -> GeometrySettings:
-        """ styling information for the timeline """
-        return self._style
-
-    @property
-    def first(self) -> datetime:
-        """ first date of the timeline """
-        return self._first
-
-    @property
-    def last(self) -> datetime:
-        """ last date of the timeline """
-        return self._last
-
-    @property
-    def width(self) -> int:
-        """ full width of the canvas """
-        return self._style.canvas.width
-
-    @property
-    def height(self) -> int:
-        """ full height of the canvas """
-        return self._style.canvas.height
-
-    @property
-    def lane_normal(self) -> Vector:
-        """ Normal vector orthogonal to the timeline direction
-        This vector is used to calculate the positions of the different lanes.
-        """
-        return (self._gradient.target - self._gradient.source).orthogonal(ccw=True)
-
-    def as_coord(self, date: datetime, lane: float = 0) -> Vector:
-        """ return the coordinates responding to this date on a given lane
-        (default: on the time arrow)
-        """
-        date_coord = self._gradient.date_to_coord(date)
-        lane_point = date_coord + lane * self._style.lane.width * self.lane_normal
-        return lane_point
 
 
 class TimeLineElement(ABC):
